requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.1.0
box2d-py==2.3.8
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
future==0.17.1
gast==0.2.0
graphviz==0.10.1
grpcio==1.16.1
gym==0.10.9
h5py==2.8.0
idna==2.7
isort==4.3.4
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.15.4
progressbar2==3.38.0
protobuf==3.6.1
pyglet==1.3.2
pylint==2.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
PyYAML==3.13
requests==2.20.1
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
